Move PKI components to PKI Folder (#15963)
* params

* fix tests

* role-pki to pki-role

* role-pki-edit to pki/role-pki-edit

* configure-pki-secret component

* config-pki and config-pki-ca components

* fix tests

* pki-cert-show and pki-cert-popup

* fix
