+++ ui/app/adapters/pki/pki-role.js
@@ -1,5 +1,5 @@
 import { assign } from '@ember/polyfills';
-import ApplicationAdapter from './application';
+import ApplicationAdapter from '../application';
 import { encodePath } from 'vault/utils/path-encoding-helpers';
 
 export default ApplicationAdapter.extend({
