+++ ui/app/models/pki-ca-certificate.js
@@ -1,6 +1,6 @@
 import { attr } from '@ember-data/model';
 import { computed } from '@ember/object';
-import Certificate from './pki-certificate';
+import Certificate from './pki/pki-certificate';
 
 export default Certificate.extend({
   DISPLAY_FIELDS: computed(function () {
