+++ ui/app/models/pki-certificate-sign.js
@@ -1,7 +1,7 @@
 import { attr } from '@ember-data/model';
 import { copy } from 'ember-copy';
 import { computed } from '@ember/object';
-import Certificate from './pki-certificate';
+import Certificate from './pki/pki-certificate';
 import { combineFieldGroups } from 'vault/utils/openapi-to-attrs';
 
 export default Certificate.extend({
