+++ ui/app/routes/vault/cluster/secrets/backend/list.js
@@ -80,7 +80,7 @@ export default Route.extend({
       ssh: 'role-ssh',
       transform: this.modelTypeForTransform(tab),
       aws: 'role-aws',
-      pki: tab === 'certs' ? 'pki-certificate' : 'role-pki',
+      pki: tab === 'certs' ? 'pki/pki-certificate' : 'pki/pki-role',
       // secret or secret-v2
       cubbyhole: 'secret',
       kv: secretEngine.get('modelTypeForKV'),
@@ -138,7 +138,7 @@ export default Route.extend({
       // possible that there is no certificate for them in order to know,
       // we fetch them specifically on the list page, and then unload the
       // records if there is no `certificate` attribute on the resultant model
-      ['ca', 'crl', 'ca_chain'].map((id) => this.store.queryRecord('pki-certificate', { id, backend }))
+      ['ca', 'crl', 'ca_chain'].map((id) => this.store.queryRecord('pki/pki-certificate', { id, backend }))
     ).then(
       (results) => {
         results.rejectBy('certificate').forEach((record) => record.unloadRecord());
