+++ ui/app/routes/vault/cluster/secrets/backend/secret-edit.js
@@ -101,7 +101,7 @@ export default Route.extend(UnloadModelRoute, {
       ssh: 'role-ssh',
       transform: this.modelTypeForTransform(secret),
       aws: 'role-aws',
-      pki: secret && secret.startsWith('cert/') ? 'pki-certificate' : 'role-pki',
+      pki: secret && secret.startsWith('cert/') ? 'pki/pki-certificate' : 'pki/pki-role',
       cubbyhole: 'secret',
       kv: backendModel.get('modelTypeForKV'),
       keymgmt: `keymgmt/${options.queryParams?.itemType || 'key'}`,
@@ -230,7 +230,7 @@ export default Route.extend(UnloadModelRoute, {
     if (!secret) {
       secret = '\u0020';
     }
-    if (modelType === 'pki-certificate') {
+    if (modelType === 'pki/pki-certificate') {
       secret = secret.replace('cert/', '');
     }
     if (modelType.startsWith('transform/')) {
