+++ ui/app/serializers/pki/pki-role.js
@@ -1,3 +1,3 @@
-import RoleSerializer from './role';
+import RoleSerializer from '../role';
 
 export default RoleSerializer.extend();
