+++ ui/app/templates/components/pki/config-pki.hbs
