+++ ui/app/templates/components/pki/configure-pki-secret.hbs
