+++ ui/app/templates/components/pki/pki-cert-show.hbs
