+++ ui/app/templates/components/secret-list/pki-cert-item.hbs
@@ -18,7 +18,7 @@
       </LinkTo>
     </div>
     <div class="column has-text-right">
-      <PkiCertPopup @item={{@item}} />
+      <Pki::PkiCertPopup @item={{@item}} />
     </div>
   </div>
 </LinkedBlock>
