+++ ui/app/templates/vault/cluster/settings/configure-secret-backend.hbs
@@ -36,7 +36,7 @@
     @saveAWSLease={{action "save" "saveAWSLease"}}
   />
 {{else if (eq this.model.type "pki")}}
-  <ConfigurePkiSecret />
+  <Pki::ConfigurePkiSecret />
 {{else if (eq this.model.type "ssh")}}
   <ConfigureSshSecret @model={{this.model}} @configured={{this.configured}} @saveConfig={{action "saveConfig"}} />
 {{/if}}
