+++ ui/app/templates/vault/cluster/settings/configure-secret-backend/section.hbs
@@ -1,7 +1,7 @@
 {{#if (eq this.model.backendType "pki")}}
   {{#if (eq this.model.section "cert")}}
-    <ConfigPkiCa @config={{this.model}} @onRefresh={{this.onRefresh}} />
+    <Pki::ConfigPkiCa @config={{this.model}} @onRefresh={{this.onRefresh}} />
   {{else}}
-    <ConfigPki @config={{this.model}} @onRefresh={{this.onRefresh}} @section={{this.model.section}} />
+    <Pki::ConfigPki @config={{this.model}} @onRefresh={{this.onRefresh}} @section={{this.model.section}} />
   {{/if}}
 {{/if}}
