+++ ui/tests/pages/components/pki/config-pki-ca.js
@@ -1,6 +1,6 @@
 import { clickable, collection, fillable, text, selectable, isPresent } from 'ember-cli-page-object';
 
-import fields from './form-field';
+import fields from '../form-field';
 
 export default {
   ...fields,
