+++ ui/tests/pages/settings/configure-secret-backends/pki/section-cert.js
@@ -1,6 +1,6 @@
 import { create, visitable } from 'ember-cli-page-object';
 
-import ConfigPKICA from 'vault/tests/pages/components/config-pki-ca';
+import ConfigPKICA from 'vault/tests/pages/components/pki/config-pki-ca';
 import flashMessages from 'vault/tests/pages/components/flash-message';
 
 export default create({
